+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 
 - Improve error message for `pd.to_snowflake`, `DataFrame.to_snowflake`, and `Series.to_snowflake` when the table does not exist.
 - Improve readability of docstring for the `if_exists` parameter in `pd.to_snowflake`, `DataFrame.to_snowflake`, and `Series.to_snowflake`.
+- Support relaxed consistency and ordering guarantees in `pd.read_snowflake` for non-query data sources.
 
 ## 1.28.0 (2025-02-20)